Switch JaCoCo on-the-fly to offline for integration tests.
Add a Java module with extra build logic to make all classes are instrumented by JaCoCo before launching the Gradle Test JVM.

Note that when using JaCoCo on-the-fly, JaCoCo instruments the classes after Robolectric. So there is nothing to control over by Robolectric.
