+++ integration_tests/jacoco-offline/src/main/java/org/robolectric/integrationtests/jacoco/JaCoCoTester.java
@@ -1,0 +1,9 @@
+package org.robolectric.integrationtests.jacoco;
+
+public class JaCoCoTester {
+  public static final int VALUE = 1;
+
+  public int getValue() {
+    return VALUE;
+  }
+}
