--- integration_tests/jacoco-on-the-fly/build.gradle
@@ -1,38 +1,0 @@
-import org.robolectric.gradle.AndroidProjectConfigPlugin
-
-apply plugin: 'com.android.library'
-apply plugin: AndroidProjectConfigPlugin
-apply plugin: "jacoco"
-
-android {
-    compileSdk 33
-
-    defaultConfig {
-        minSdk 16
-        targetSdk 33
-    }
-
-    compileOptions {
-        sourceCompatibility = '1.8'
-        targetCompatibility = '1.8'
-    }
-
-    testOptions.unitTests.includeAndroidResources true
-}
-
-jacoco {
-    toolVersion = jacocoVersion
-}
-
-dependencies {
-    compileOnly AndroidSdk.MAX_SDK.coordinates
-
-    testCompileOnly AndroidSdk.MAX_SDK.coordinates
-    testRuntimeOnly AndroidSdk.MAX_SDK.coordinates
-
-    testImplementation project(":robolectric")
-    testImplementation "junit:junit:$junitVersion"
-    testImplementation "com.google.truth:truth:$truthVersion"
-    testImplementation "androidx.test.ext:junit:$axtJunitVersion"
-    testImplementation "org.jacoco:org.jacoco.agent:$jacocoVersion:runtime"
-}
