--- integration_tests/jacoco-on-the-fly/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.robolectric.integrationtests.jacoco.onthefly">
-</manifest>
