--- integration_tests/jacoco-on-the-fly/src/main/java/org/robolectric/integrationtests/jacoco/onthefly/JaCoCoOnTheFlyTester.java
@@ -1,9 +1,0 @@
-package org.robolectric.integrationtests.jacoco.onthefly;
-
-public class JaCoCoOnTheFlyTester {
-  public static final int VALUE = 1;
-
-  public int getValue() {
-    return VALUE;
-  }
-}
